Fixed the CDCL algorithm, implemented EVSIDS and changed how the two watched literals are accessed
The CDCL algorithm didn't use the first UIP for clause learning. This should now be fixed.
The two first literals of a clause are now always the watched literals. If one of the watched
literals turn false, it swaps positions with another none false literal. I also
implemented the EVSIDS variable selection strategy.